+++ package.json
@@ -18,6 +18,7 @@
     "dotenv": "16.3.1",
     "express": "4.18.2",
     "jsonwebtoken": "9.0.2",
+    "morgan": "^1.10.0",
     "mysql2": "3.6.1",
     "nodemailer": "6.9.7",
     "oracledb": "6.2.0",
