+++ src/index.js
@@ -1,21 +1,33 @@
 import { chatBootClient } from './routes/chatBoot.routes.js'
 import { routerCF } from './routes/clienteFiel.routes.js'
-import { corsMiddleware } from '../middlewares/cors.js'
 import { routerUser } from './routes/users.routes.js'
 import cookieParser from 'cookie-parser'
+import morgan from 'morgan'
 import express from 'express'
 import dotenv from 'dotenv'
+import cors from 'cors'
 
 dotenv.config()
 
+const ACCEPTED_ORIGINS = [
+  'http://172.20.1.160'
+]
+
 const app = express()
+
+app.disable('x-powered-by')
 const PORT = process.env.PUERTO_API || 4000
+
+app.use(cors({
+  credentials: true,
+  origin: ACCEPTED_ORIGINS
+}))
+
+app.use(morgan('dev'))
 
 app.use(cookieParser())
 app.use(express.json())
 
-app.disable('x-powered-by')
-app.use(corsMiddleware())
 
 // TODO: Metodos En Usuarios Login
 app.use(routerUser)
